refactor messagebar usage, now only on front page
