+++ app/component/splash/Splash.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import { intlShape, FormattedMessage } from 'react-intl';
-import FrontPagePanelContainer from '../front-page/FrontPagePanelContainer';
 import { setOriginToDefault } from '../../action/EndpointActions';
 import FakeSearchBar from '../search/FakeSearchBar';
 import { default as FakeSearchWithButton } from '../search/FakeSearchWithButton';
@@ -55,7 +54,6 @@ class Splash extends React.Component {
           <div className="map fullscreen">
             <div className="background-gradient" />
           </div>
-          <FrontPagePanelContainer />
         </div>
       </div>
       <div className="splash">
