+++ docs/ZIndex.md
@@ -12,6 +12,7 @@ N/A | React Material Drawer | 1200/1300 | Defined by material-ui
 `.splash` | Splash screen | 1100 |
 `div.leaflet-marker-icon.from, div.leaflet-marker-icon.to { > span { &::before` | From/To marker letters | 1000 | Could be removed through new icon components
 `.toggle-positioning-container` | Pan-to-your-position button | 802 |
+`.message-bar` | Messagebar | 802 |
 `.search-form-map-overlay` | Fake search field | 802 |
 `.fullscreen-toggle` | Map fullscreen toggle | 802 |
 `.background-gradient` | Shadow at top of map | 801 |
